+++ scripts/deployVerificationRegistryGM.ts
@@ -3,7 +3,7 @@ import { formatBytes32String } from "ethers/lib/utils";
 
 async function main() {
   const accounts = lacchain.getSigners();
-  const artifactName = "VerificationRegistry";
+  const artifactName = "VerificationRegistryGM";
   const defaultDidRegistry = "0x43dE0954a2c83A415d82b9F31705B969b5856003"; // UPDATE this!
   const defaultDelegateType = formatBytes32String("sigAuth"); // bytes32 right padded
   console.log(defaultDelegateType);