+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.0.0
+
+- Set Contract License to Apache License, Version 2.0
+- Creates Generic version of Verification Registry Smart contracts (Valid for any generic Ethereum Based Network)
+- Improves documentation
+
 ## 1.0.0
 
 - first version of verification registry contracts